% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_mds.R
\name{rf_mds}
\alias{rf_mds}
\title{rf_mds generates a two-dimensional MDS plot given random forest proximities}
\usage{
rf_mds(
  x = NULL,
  y = NULL,
  n_dim = 2,
  prox = NULL,
  rf = NULL,
  type = "rfgap",
  mds_type = "nonmetric",
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{A dataframe or matrix of data. Rows (n) are observations,
columns (d) are variables.}

\item{y}{The labels corresponding to x. Should be of type factor or numeric.}

\item{n_dim}{how many dimensions should be in the resulting mds embedding?}

\item{prox}{A pre-computed rf_proximities object. If provided, x, rf, and
type will be ignored}

\item{rf}{A forest object of s3 class 'ranger'. Note: rf requires
write.forest = TRUE and keep.inbag = TRUE.}

\item{type}{The type of proximities to be obtained.
Options are rfgap (default), original, oob, pbk, or rfproxih.}

\item{mds_type}{Type of MDS to be run; 'metric' (default) or 'nonmetric'}

\item{seed}{random state for the random forest}

\item{...}{additional ranger options.  Not used if rf is supplied}
}
\value{
MDS coordinates. A plot will be generated if plot = TRUE
}
\description{
rf_mds generates a two-dimensional MDS plot given random forest proximities
}
