% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_upsample.R
\name{rf_upsample}
\alias{rf_upsample}
\title{This function generates random forest proximities of a variety of types.
THIS FUNCTION IS NOT CURRENLY IMPLEMENTED}
\usage{
rf_upsample(
  x,
  y,
  rf = NULL,
  type = "rfgap",
  x_test = NULL,
  prox = NULL,
  n_samples = 2 * nrow(x),
  class_weights = "uniform",
  ...
)
}
\arguments{
\item{x}{A dataframe or matrix of data. Rows (n) are observations,
columns (d) are variables.}

\item{y}{The labels corresponding to x. If categorical, should be of type
'factor'.}

\item{rf}{A forest object of s3 class 'ranger'. Note: rf requires
write.forest = TRUE and keep.inbag = TRUE.}

\item{type}{The type of proximities to be obtained.
Options are rfgap (default), original, oob, pbk, or rfproxih.}

\item{x_test}{An (optional) test dataset (dataframe or matrix)}

\item{prox}{a precomputed random forest proximity matrix}

\item{n_samples}{How many samples should be generated?}

\item{class_weights}{how the number of samples should be weighted. Currently,
only 'uniform' is supported}

\item{...}{Additional argument options for ranger(). (If rf is not supplied).}
}
\value{
A matrix of proximity values.
}
\description{
This function generates random forest proximities of a variety of types.
THIS FUNCTION IS NOT CURRENLY IMPLEMENTED
}
\examples{

x <- iris[, -5]
y <- iris[, 5]

}
